+++ MORE/CustomSay.cs
@@ -1,16 +1,12 @@
-﻿using System.Collections.Generic;
-
-namespace Shockah.MORE;
+﻿namespace Shockah.MORE;
 
 internal sealed class CustomSay : Say
 {
 	private static int NextId = 1;
 
 	private readonly int Id = 0;
-	private int ExecuteCount = 0;
 
 	public string? Text;
-	public List<string>? AlternativeTexts;
 
 	public CustomSay()
 	{
@@ -19,16 +15,11 @@ internal sealed class CustomSay : Say
 
 	public override bool Execute(G g, IScriptTarget target, ScriptCtx ctx)
 	{
-		var text = Text;
-		if (string.IsNullOrEmpty(text) && AlternativeTexts is not null && AlternativeTexts.Count != 0)
-			text = AlternativeTexts[ExecuteCount % AlternativeTexts.Count];
-
-		if (string.IsNullOrEmpty(text))
+		if (string.IsNullOrEmpty(Text))
 			return base.Execute(g, target, ctx);
 
 		hash = $"{GetType().FullName}:{Id}";
-		DB.currentLocale.strings[GetLocKey(ctx.script, hash)] = text;
-		ExecuteCount++;
+		DB.currentLocale.strings[GetLocKey(ctx.script, hash)] = Text;
 		return base.Execute(g, target, ctx);
 	}
 }