+++ MORE/docs/release-notes.md
@@ -2,6 +2,11 @@
 
 # Release notes
 
+## Upcoming release
+
+* The Ship Swap event now re-applies artifacts which do ship-wide modifications. Other mods can use an API to add more artifacts to this system.
+* Fixed a mistake in The Duplicates Trial's description.
+
 ## 1.0.1
 Released 7 July 2024.